chore(web-components): delete NextJS storybook
I've recently added NextJS to my no-no list of frameworks I avoid. I'm currently planning to keep web-components-wrapper-nextjs but that could change on a dime.

I have wasted a lot of time fighting against NextJS instead of working with it. The over-the-top caching, unstable app directory, overriding of fetch, inability to tree-shake barrel files, slow "fast refresh", redirects as exceptions, and breaking changes on minor upgrades are all too much for me to handle. And these are just my complaints off the dome.