refactor(web-components): align css reset, simplify IconContainer
I have always planned to use a consistent stylesheet between all my projects. I realised today that ought to include the web-components Storybook. So I'm pulling in more of those styles, allowing me to simplify IconContainer as it does not have to cater to two stylesheets. These components are prefixed with my name for a reason.